+++ sn_camels/models/sn_base_models.py
@@ -19,9 +19,10 @@ import cv2
 
 import torch.nn as nn
 
-from kymatio.scattering2d.frontend.torch_frontend import ScatteringTorch2D
+from kymatio import Scattering2D
 
 from sn_camels.scattering.create_filters import *
+from sn_camels.scattering.scattering2d import construct_scattering
 from sn_camels.models.models_utils import get_filters_visualization, getOneFilter, getAllFilters,compareParams, compareParamsVisualization
 
 
@@ -46,7 +47,7 @@ def create_scatteringExclusive(J,N,M,second_order,device,initialization,seed=0,r
     seed -- the seed used for creating randomly initialized filters
     requires_grad -- boolean idicating whether we want to learn params
     """
-    scattering = ScatteringTorch2D(J=J, shape=(M, N))
+    scattering = Scattering2D(J=J, shape=(M, N), frontend='torch')
 
     L = scattering.L
     if second_order:
